+++ src/main/java/Krypto/Commands/Command.java
@@ -1,3 +1,7 @@
+package Krypto.Commands;
+import Krypto.IO.*;
+import Krypto.Utils.*;
+
 public abstract class Command {
     public abstract void execute(UI ui, TaskList tasks, Storage storage);
     public abstract boolean isExit();
