+++ src/main/java/Krypto/Commands/DeleteCommand.java
@@ -1,10 +1,13 @@
-public class UnmarkCommand extends Command {
+package Krypto.Commands;
+import Krypto.IO.*;
+import Krypto.Utils.*;
+public class DeleteCommand extends Command {
     private int index;
-    public UnmarkCommand(int index) {
+    public DeleteCommand(int index) {
         this.index = index;
     }
     public void execute(UI ui, TaskList tasks, Storage storage) {
-        tasks.getTask(index).unmarkTask();
+        tasks.deleteTask(index);
         storage.store(tasks);
     }
     public  boolean isExit() {
