+++ src/main/java/Krypto/Commands/UnmarkCommand.java
@@ -1,10 +1,13 @@
-public class DeleteCommand extends Command {
+package Krypto.Commands;
+import Krypto.IO.*;
+import Krypto.Utils.*;
+public class UnmarkCommand extends Command {
     private int index;
-    public DeleteCommand(int index) {
+    public UnmarkCommand(int index) {
         this.index = index;
     }
     public void execute(UI ui, TaskList tasks, Storage storage) {
-        tasks.deleteTask(index);
+        tasks.getTask(index).unmarkTask();
         storage.store(tasks);
     }
     public  boolean isExit() {
