+++ src/main/java/Krypto/Exceptions/KryptoExceptions.java
@@ -1,3 +1,4 @@
+package Krypto.Exceptions;
 public class KryptoExceptions extends Exception{
     public KryptoExceptions(String message) {
         super(message);
