+++ src/main/java/Krypto/Exceptions/ToDoException.java
@@ -1,3 +1,4 @@
+package Krypto.Exceptions;
 public class ToDoException extends KryptoExceptions {
 
     public ToDoException() {
