+++ src/main/java/Krypto/Task/Deadline.java
@@ -1,3 +1,4 @@
+package Krypto.Task;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
