+++ src/main/java/Krypto/Task/Event.java
@@ -1,6 +1,6 @@
+package Krypto.Task;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
-import java.time.LocalDateTime;
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
