+++ src/main/java/Krypto/Task/Task.java
@@ -1,3 +1,4 @@
+package Krypto.Task;
 public class Task {
     protected String description;
     protected boolean isDone = false;
