+++ src/main/java/Krypto/Task/ToDo.java
@@ -1,8 +1,7 @@
-import java.time.LocalDate;
+package Krypto.Task;
+
 
 public class ToDo extends Task {
-
-
     public ToDo(String description) {
         super(extractContent(description.split("/")[0].split(" ")));
     }
@@ -15,6 +14,5 @@ public class ToDo extends Task {
     public String toFileString() {
         return "T | " + (super.getStatusIcon()) + " | " + description;
     }
-
 
 }
